langgraph-agents-demo/requirements.txt
# Core LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-anthropic>=0.2.0
langchain-community==0.3.5
langchain-core==0.3.15

# Web Interface
gradio==4.44.0

# External APIs
tavily-python==0.5.0
langchain-tavily>=0.2.0  # Updated Tavily integration for LangChain
amadeus==9.1.0

# Utilities
python-dotenv==1.0.0

# Image processing for graph visualization
pillow==10.4.0
